# torque-api-spec.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Torque API Reference",
    "description": "This page contains information about the Torque APIs and how to use it.<br><br>To be able to test the API methods, you will need an access token to be set in the Authorize section.If you got to this page from your Torque account, we already made it available for you, but you can switch to a different token as needed.<br>To get a different token, use the 'Get token' function available under the 'Access Tokens' section. After running the method, copy the access_token you get in the response and set it in the Authorize fieldas 'Bearer access_token'. For example: Bearer fqSWuw72BbUVFn8AbokF77GJ0r5KEn9MiZjLXF8kBwI.<br><br>The API can be accessed with your account endpoint as will be mentioned in the descriptions and examples or using https://qtorque.io as the prefix instead."
  },
  "paths": {
...
    "/api/spaces/{space_name}/sandbox": {
      "get": {
        "tags": [
          "Sandbox Environments"
        ],
        "summary": "Get all sandboxes",
        "description": "Returns details of all sandboxes in the space.",
        "parameters": [
          {
            "name": "space_name",
            "in": "path",
            "description": "Name of the space containing the sandboxes",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "filter",
            "in": "query",
            "description": "Filter for the environment:\r\n- \"automation\": Sandboxes launched by any automation process in the space\r\n- \"my\": Sandboxes launched by the user\r\n- \"all\": Sandboxes launched by anyone in the space",
            "schema": {
              "type": "string",
              "default": "all"
            }
          },
          {
            "name": "sandbox_name",
            "in": "query",
            "description": "String that all sandbox names must contain",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "count",
            "in": "query",
            "description": "Amount of sandboxes to be retrieved",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 20
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SandboxResponseListItem"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Space not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Count is above the maximum allowed sandboxes (1000)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Sandbox Environments"
        ],
        "summary": "Start new sandbox",
        "description": "Launches a new sandbox environment in Torque",
        "parameters": [
          {
            "name": "space_name",
            "in": "path",
            "description": "Name of the space that should contain the sandbox environment",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "description": "",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSandboxRequest"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateEnvResponse"
                }
              }
            }
          },
          "400": {
            "description": "Operation failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "User is not permitted to launch sandbox from branch",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Space, current user, or compute service user not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Could not processed request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "424": {
            "description": "Cloud account not accessible",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/spaces/{space_name}/sandbox/{sandbox_id}": {
      "get": {
        "tags": [
          "Sandbox Environments"
        ],
        "summary": "Gets sandbox details",
        "description": "Gets details on a sandbox.",
        "parameters": [
          {
            "name": "space_name",
            "in": "path",
            "description": "Name of the space containing the sandbox",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sandbox_id",
            "in": "path",
            "description": "Sandbox ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SandboxResponse"
                }
              }
            }
          },
          "404": {
            "description": "Space or sandbox not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "424": {
            "description": "Cloud account is not accessible",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Sandbox Environments"
        ],
        "summary": "End sandbox",
        "description": "Ends a running sandbox from Torque.",
        "parameters": [
          {
            "name": "space_name",
            "in": "path",
            "description": "Name of the space containing the sandbox",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sandbox_id",
            "in": "path",
            "description": "Sandbox ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "ignore_errors",
            "in": "query",
            "description": "Should ignore errors during termination, if set to true then if termination of the sandbox failed it will be accepted as termination succeeded.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "202": {
            "description": "Success"
          },
          "400": {
            "description": "Operation failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Space not found, sandbox not found, or sandbox already ended",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Invalid \"ignore_errors\" value",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "423": {
            "description": "Sandbox already ending or temporarily unavailable for ending",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "424": {
            "description": "Cloud not accessible",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/spaces/{space_name}/sandbox/{sandbox_id}/debuggingservice": {
      "put": {
        "tags": [
          "Sandbox Environments"
        ],
        "summary": "Turn Bastion on or off",
        "description": "This API requires Bastion to be deployed in the sandbox environment",
        "parameters": [
          {
            "name": "space_name",
            "in": "path",
            "description": "Name of the space containing the sandbox",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sandbox_id",
            "in": "path",
            "description": "Sandbox ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "value",
            "in": "query",
            "description": "Indicates if to enable/disable the debugging service: \"on\" or \"off\"",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "202": {
            "description": "Request is in progress or triggered"
          },
          "400": {
            "description": "Operation failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Space not found, sandbox not found, or debugging service is disabled",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "The current status of the debugging service is conflicted with the new status request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "\"value\" is empty or invalid",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "424": {
            "description": "Cloud not accessible",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/spaces/{space_name}/sandbox/{sandbox_id}/scheduled_end_time": {
      "put": {
        "tags": [
          "Sandbox Environments"
        ],
        "summary": "Extend sandbox",
        "description": "Extends the duration of a running sandbox",
        "parameters": [
          {
            "name": "space_name",
            "in": "path",
            "description": "Name of the space containing the sandbox",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sandbox_id",
            "in": "path",
            "description": "Sandbox ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "value",
            "in": "query",
            "description": "New scheduled end time in ISO 8601 format. For example, 2021-10-06T08:27:05.215Z.",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "Operation failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Space or sandbox not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "\"value\" (end time) is empty, not UTC, or time already passed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "424": {
            "description": "Cloud not accessible",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/spaces/{space_name}/sandbox/{sandboxId}/logs/tfapply": {
      "get": {
        "tags": [
          "Sandbox Environments"
        ],
        "summary": "Get terraform apply logs",
        "parameters": [
          {
            "name": "space_name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sandboxId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "service_name",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/spaces/{space_name}/sandbox/{sandboxId}/logs/tfdestroy": {
      "get": {
        "tags": [
          "Sandbox Environments"
        ],
        "summary": "Get terraform destroy logs",
        "parameters": [
          {
            "name": "space_name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sandboxId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "service_name",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreateEnvResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Environment ID",
            "nullable": true
          }
        },
        "additionalProperties": false,
        "description": "Create environment response",
        "x-internal": true
      },
...
      "Error": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "description": "Error message",
            "nullable": true
          },
          "name": {
            "type": "string",
            "description": "Error name",
            "nullable": true
          },
          "code": {
            "type": "string",
            "description": "Error code",
            "nullable": true
          }
        },
        "additionalProperties": false,
        "description": "Error information",
        "x-internal": true
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "errors": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Error"
            },
            "description": "Array of errors",
            "nullable": true
          }
        },
        "additionalProperties": false,
        "description": "Errors received from an API",
        "x-internal": true
      },
...
    },
    "securitySchemes": {
      "Bearer": {
        "type": "apiKey",
        "description": "Authorization header using the Bearer scheme. Example: \"Authorization: Bearer {token}\"",
        "name": "Authorization",
        "in": "header"
      }
    }
  },
  "security": [
    {
      "Bearer": [ ]
    }
  ]
}